+++ source/extensions/isaacsim.asset.importer.mjcf.ui/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.1] - 2026-02-26
+### Changed
+- Remove extension api doc
+
 ## [1.0.0] - 2026-01-01
 ### Changed
 - Decoupled UI from the MJCF importer